adapter/requirements.txt
mysql-connector
requests
influxdb
